requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
lxml
anthropic==0.40.0
python-dotenv==1.0.0
flask==3.0.0
supabase==2.3.4
gunicorn==21.2.0
